add event lister to the button components by props, and all buttons use the function and i used event.target.name to know which button is clicked and send the calculator state data with the button clicked to the logic calculate fucntion and then display the data in the calculator for user.
