+++ src/components/input.jsx
@@ -1,7 +1,22 @@
 import React, { Component } from 'react';
-import './input_style.css'
+import './input_style.css';
 
 class Input extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  getCalculateData = () => {
+    const { total, next, operation } = this.props.result;
+    if (total && next && operation) return `${total} ${operation} ${next}`;
+    if (total && operation) return `${total} ${operation} `;
+    if (total && !next) return `${total}`;
+    if (!total && next) return `${next}`;
+
+    // if there is no data like total and next return 0
+    return 0;
+  };
+
   render() {
     return (
       <div className="input_container">
@@ -10,7 +25,8 @@ class Input extends Component {
           type="text"
           name="number"
           id="number"
-          value={0}
+          disabled
+          value={this.getCalculateData()}
         />
       </div>
     );
